requirements-web.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# File uploads
python-multipart>=0.0.6
aiofiles>=23.2.1

# WebSockets
websockets>=12.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0  # MySQL driver
psycopg2-binary>=2.9.0  # PostgreSQL driver (optional)

# AWS S3 (for cloud deployment)
boto3>=1.34.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Settings management
pydantic-settings>=2.0.0

# Email validation
email-validator>=2.0.0

# Existing dependencies (from badminton analyzer)
# opencv-python>=4.8.0
# mediapipe>=0.10.0
# numpy>=1.24.0
